+++ cmd/neofs-ir/defaults_test.go
@@ -30,10 +30,13 @@ func TestValidateDefaultConfig(t *testing.T) {
 			Validators:        keys.PublicKeys{},
 		},
 		NNS: config.NNS{SystemEmail: ""},
-		Mainnet: config.BasicChain{
-			DialTimeout:         60000000000,
-			ReconnectionsNumber: 5,
-			ReconnectionsDelay:  5000000000,
+		Mainnet: config.Mainnet{
+			Enabled: false,
+			BasicChain: config.BasicChain{
+				DialTimeout:         60000000000,
+				ReconnectionsNumber: 5,
+				ReconnectionsDelay:  5000000000,
+			},
 		},
 		Control: config.Control{
 			AuthorizedKeys: keys.PublicKeys{},
@@ -45,7 +48,6 @@ func TestValidateDefaultConfig(t *testing.T) {
 				Path: ".neofs-ir-state",
 			},
 		},
-		Fee: config.Fee{MainChain: 50000000},
 		Timers: config.Timers{
 			CollectBasicIncome: config.BasicTimer{Mul: 1, Div: 2},
 		},