+++ CHANGELOG.md
@@ -8,12 +8,17 @@ Changelog for NeoFS Node
 ### Fixed
 
 ### Changed
+- Move `fschain_autodeploy` into `fschain.disable_autodeploy` in IR config (#3619)
 
 ### Removed
 
 ### Updated
 
 ### Updating from v0.49.0
+Use IR configuration option:
+- `fschain.disable_autodeploy` instead of deprecated `fschain_autodeploy` with reverted value,
+by default autodeploy is enabled now;
+Old option is still supported but will be removed in future releases.
 
 ## [0.49.0] - 2025-10-06 - Dochodo